backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
google-cloud-aiplatform==1.38.1
google-generativeai==0.3.0
redis==5.0.1
python-dotenv==1.0.0
Pillow==10.0.0
aiohttp==3.9.0
numpy==1.26.4
